% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{arc}
\alias{arc}
\title{Draw an Arc}
\usage{
arc(sketch, a = NULL, b = NULL, c = NULL, d = NULL, start = NULL,
  stop = NULL, mode = NULL)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{a}{The x coordinate of the arc's ellipse.}

\item{b}{The y coordinate of the arc's ellipse.}

\item{c}{The width of the arc's ellipse.}

\item{d}{The height of the arc's ellipse.}

\item{start}{The angle to start the arc in radians.}

\item{stop}{The angle to stop the arc in radians.}

\item{mode}{Determines the style by which the arc is drawn.}
}
\description{
Draw an Arc
}
\examples{
\dontrun{

p5() \%>\%
  createCanvas(100, 100) \%>\%
  background("#DCDCDC") \%>\%
  arc(50, 55, 50, 50, 0, ~HALF_PI) \%>\%
  noFill() \%>\%
  arc(50, 55, 60, 60, ~HALF_PI, ~PI) \%>\%
  arc(50, 55, 70, 70, ~PI, ~PI+QUARTER_PI) \%>\%
  arc(50, 55, 80, 80, ~PI+QUARTER_PI, ~TWO_PI)

p5() \%>\%
  background("#DCDCDC") \%>\%
  arc(50, 50, 80, 80, 0, ~PI+QUARTER_PI, ~OPEN)

p5() \%>\%
  background("#DCDCDC") \%>\%
  arc(50, 50, 80, 80, 0, ~PI+QUARTER_PI, ~CHORD)

p5() \%>\%
  background("#DCDCDC") \%>\%
  arc(50, 50, 80, 80, 0, ~PI+QUARTER_PI, ~PIE)

}
}
\seealso{
Other Shape 2D Primitives: \code{\link{ellipse}},
  \code{\link{line}}, \code{\link{rect}}
}
