% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{background}
\alias{background}
\title{Set the canvas bachground color}
\usage{
background(sketch, colorstring)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{colorstring}{A hex color string.}
}
\description{
Set the canvas bachground color
}
\examples{
\dontrun{

p5() \%>\%
  background("#FAE")

setup() \%>\%
  background(rgb(255, 204, 0, maxColorValue = 255))

draw() \%>\%
  background("#808080")

}
}
