% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{between}
\alias{between}
\title{Make a between sketch}
\usage{
between(data = NULL)
}
\arguments{
\item{data}{An optional \code{\link[tibble:data_frame]{tibble::data_frame()}}.}
}
\description{
Make a between sketch
}
