% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{bind_sketches}
\alias{bind_sketches}
\title{Combine multiple sketches together}
\usage{
bind_sketches(..., width = NULL, height = NULL, padding = 0)
}
\arguments{
\item{...}{A series of p5 sketches.}

\item{width}{Width of the sketch.}

\item{height}{Height of the sketch.}

\item{padding}{Padding of the sketch.}
}
\description{
Combine multiple sketches together
}
\examples{
\dontrun{

library(dplyr)

stripes <- data_frame(
  x = rep(0, 7),
  y = cumsum(c(0, rep(30, 6))),
  w = rep(190*2, 7),
  h = rep(15, 7)
)

stripes_ <- stripes \%>\%
  p5() \%>\%
  createCanvas(190*2, 200) \%>\%
  fill("#B22234") \%>\%
  noStroke() \%>\%
  rect()

stars_ <- p5() \%>\%
  fill("#3C3B6E") \%>\%
  noStroke() \%>\%
  rect(0, 0, 152, 105)

bind_sketches(stripes_, stars_)

}
}
