% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{createCanvas}
\alias{createCanvas}
\title{Set the dimensions of the sketch canvas}
\usage{
createCanvas(sketch, w, h, renderer = NULL)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{w}{The width of the canvas in pixels.}

\item{h}{The height of the canvas in pixels.}

\item{renderer}{Either the default renderer (\code{NULL}) or \code{~WEBGL}.}
}
\description{
Set the dimensions of the sketch canvas
}
\examples{
\dontrun{

p5() \%>\%
  createCanvas(300, 200) \%>\%
  background("#FAE")

}
}
