% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{ellipse}
\alias{ellipse}
\title{Draw an Ellipse}
\usage{
ellipse(sketch, x = NULL, y = NULL, w = NULL, h = NULL)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{x}{The x coordinate of the ellipse.}

\item{y}{The y coordinate of the ellipse.}

\item{w}{The width of the ellipse.}

\item{h}{The height of the ellipse.}
}
\description{
Draw an Ellipse
}
\examples{
\dontrun{

p5() \%>\%
  background("#DCDCDC") \%>\%
  ellipse(56, 46, 55, 55)

}
}
\seealso{
Other Shape 2D Primitives: \code{\link{arc}},
  \code{\link{line}}, \code{\link{rect}}
}
