% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{fill}
\alias{fill}
\title{Set shape color}
\usage{
fill(sketch, colorstring)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{colorstring}{A hex color string.}
}
\description{
Set shape color
}
\examples{
\dontrun{

p5() \%>\%
  fill("#FAE") \%>\%
  rect(10, 10, 10, 10) \%>\%
  fill("#808080") \%>\%
  rect(30, 30, 10, 10)

}
}
