% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{js}
\alias{js}
\title{Add Javascript to any part of a sketch}
\usage{
js(sketch, ..., objn = "p.")
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{...}{Strings of Javascript code to add to the sketch.}

\item{objn}{The name of the sketch object followed by a period. Most of the
time you should keep the default \code{p.}, however it can be useful to change this
to and empty string (\code{""}).}
}
\description{
Add Javascript to any part of a sketch
}
