% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{line}
\alias{line}
\title{Draw a Line}
\usage{
line(sketch, x1 = NULL, y1 = NULL, x2 = NULL, y2 = NULL)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{x1}{The x coordinate for the first point.}

\item{y1}{The y coordinate for the first point.}

\item{x2}{The x coordinate for the second point.}

\item{y2}{The y coordinate for the second point.}
}
\description{
Draw a Line
}
\examples{
\dontrun{

p5() \%>\%
  line(30, 20, 85, 75)

}
}
\seealso{
Other Shape 2D Primitives: \code{\link{arc}},
  \code{\link{ellipse}}, \code{\link{rect}}
}
