% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{noFill}
\alias{noFill}
\title{Disable shape color}
\usage{
noFill(sketch)
}
\arguments{
\item{sketch}{A p5 sketch.}
}
\description{
Disable shape color
}
\examples{
\dontrun{

p5() \%>\%
  fill("#FAE") \%>\%
  rect(10, 10, 10, 10) \%>\%
  noFill() \%>\%
  rect(30, 30, 10, 10)

}
}
