% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{noLoop}
\alias{noLoop}
\title{Stop draw() from looping.}
\usage{
noLoop(sketch)
}
\arguments{
\item{sketch}{A p5 sketch.}
}
\description{
Stop draw() from looping.
}
\examples{
\dontrun{
# Click the resulting sketch

pre_ <- pre() \%>\%
  js("var x = 0;")
setup_ <- setup() \%>\%
  createCanvas(100, 100)
draw_ <- draw() \%>\%
  background("#F0F8FF") \%>\%
  js(" x = x + 0.1;
       if (x > width) {
         x = 0;
       } ") \%>\%
   line(~x, 0, ~x, ~height)
post_ <- post() \%>\%
  js(" mousePressed = function() {
         noLoop();
       }

       mouseReleased = function() {
         loop();
       } ")
sketch(pre = pre_, setup = setup_, draw = draw_, post = post_)
}
}
