% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{noStroke}
\alias{noStroke}
\title{Turn off shape outlines}
\usage{
noStroke(sketch)
}
\arguments{
\item{sketch}{A p5 sketch.}
}
\description{
Turn off shape outlines
}
\examples{
\dontrun{

p5() \%>\%
  stroke("#FF0000") \%>\%
  fill("#DCDCDC") \%>\%
  rect(10, 10, 20, 30) \%>\%
  noStroke() \%>\%
  rect(40, 10, 20, 30)

}
}
