% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p5.R
\name{p5}
\alias{p5}
\title{Create a p5 sketch}
\usage{
p5(data = NULL, width = NULL, height = NULL, padding = 0)
}
\arguments{
\item{data}{A \code{\link[tibble:data_frame]{tibble::data_frame()}} which can be used to draw objects on the
canvas where each row is an object to be drawn and columns are attributes of
the object to be drawn.}

\item{width}{Width of the sketch.}

\item{height}{Height of the sketch.}

\item{padding}{Padding of the sketch.}
}
\description{
Create a blank \href{https://p5js.org/}{p5} sketch.
}
\examples{
\dontrun{

p5() \%>\%
  rect(10, 10, 40, 50)

p5() \%>\%
  createCanvas(400, 300) \%>\%
  background("#F0F8FF")

library(dplyr)

stripes <- data_frame(
  x = rep(0, 7),
  y = cumsum(c(0, rep(30, 6))),
  w = rep(300, 7),
  h = rep(15, 7)
)

stripes \%>\%
  p5() \%>\%
  createCanvas(300, 200) \%>\%
  fill("#FF0000") \%>\%
  noStroke() \%>\%
  rect()

}
}
