% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{point}
\alias{point}
\title{Draw a Point}
\usage{
point(sketch, x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{x}{The x coordinate.}

\item{y}{The y coordinate.}

\item{z}{The z coordinate (for 3D WEBGL mode).}
}
\description{
Draw a Point
}
\examples{
\dontrun{

p5() \%>\%
  point(30, 30) \%>\%
  point(40, 40) \%>\%
  point(50, 50)

(Formaldehyde * 100) \%>\%
  p5() \%>\%
  point(~carb, ~optden)

}
}
\seealso{
Other Shape 2D Primitives: \code{\link{arc}},
  \code{\link{ellipse}}, \code{\link{line}},
  \code{\link{quad}}, \code{\link{rect}},
  \code{\link{triangle}}
}
