% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{quad}
\alias{quad}
\title{Draw a Quadrilateral}
\usage{
quad(sketch, x1 = x1, y1 = y1, x2 = x2, y2 = y2, x3 = x3, y3 = y3,
  x4 = x4, y4 = y4)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{x1}{The x coordinate for the first point.}

\item{y1}{The y coordinate for the first point.}

\item{x2}{The x coordinate for the second point.}

\item{y2}{The y coordinate for the second point.}

\item{x3}{The x coordinate for the third point.}

\item{y3}{The y coordinate for the third point.}

\item{x4}{The x coordinate for the fourth point.}

\item{y4}{The y coordinate for the fourth point.}
}
\description{
Draw a Quadrilateral
}
\examples{
\dontrun{

p5() \%>\%
  quad(40, 30, 85, 25, 70, 65, 35, 75)

}
}
\seealso{
Other Shape 2D Primitives: \code{\link{arc}},
  \code{\link{ellipse}}, \code{\link{line}},
  \code{\link{point}}, \code{\link{rect}},
  \code{\link{triangle}}
}
