% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{rect}
\alias{rect}
\title{Draw a Rectangle}
\usage{
rect(sketch, x = NULL, y = NULL, w = NULL, h = NULL)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{x}{The x coordinate for the}

\item{y}{The y coordinate for the}

\item{w}{The width of the rectangle.}

\item{h}{The height of the rectangle.}
}
\description{
Draw a Rectangle
}
\examples{
\dontrun{

p5() \%>\%
  rect(10, 10, 30, 40)

library(dplyr)

data_frame(x = c(10, 30, 10, 30),
           y = c(10, 10, 30, 30),
           w = rep(10, 4),
           h = rep(10, 4)) \%>\%
  p5() \%>\%
  rect()

data_frame(A = c(10, 40),
           B = c(10, 10),
           Width = c(20, 10),
           Height = c(30, 40)) \%>\%
  p5() \%>\%
  rect(~A, ~B, ~Width, ~Height)

}
}
\seealso{
Other Shape 2D Primitives: \code{\link{arc}},
  \code{\link{ellipse}}, \code{\link{line}},
  \code{\link{point}}, \code{\link{quad}},
  \code{\link{triangle}}
}
