% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{setup}
\alias{setup}
\title{Make a setup sketch}
\usage{
setup(data = NULL)
}
\arguments{
\item{data}{An optional \code{\link[tibble:data_frame]{tibble::data_frame()}}.}
}
\description{
Make a setup sketch
}
