% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{stroke}
\alias{stroke}
\title{Set the color used to draw lines around shapes}
\usage{
stroke(sketch, colorstring)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{colorstring}{A hex color string.}
}
\description{
Set the color used to draw lines around shapes
}
\examples{
\dontrun{

p5() \%>\%
  stroke("#FF0000") \%>\%
  rect(10, 10, 20, 30) \%>\%
  stroke("#0000FF") \%>\%
  rect(20, 30, 30, 20)

}
}
