% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{triangle}
\alias{triangle}
\title{Draw a Triangle}
\usage{
triangle(sketch, x1 = x1, y1 = y1, x2 = x2, y2 = y2, x3 = x3,
  y3 = y3)
}
\arguments{
\item{sketch}{A p5 sketch.}

\item{x1}{The x coordinate for the first point.}

\item{y1}{The y coordinate for the first point.}

\item{x2}{The x coordinate for the second point.}

\item{y2}{The y coordinate for the second point.}

\item{x3}{The x coordinate for the third point.}

\item{y3}{The y coordinate for the third point.}
}
\description{
Draw a Triangle
}
\examples{
\dontrun{

p5() \%>\%
  triangle(20, 80, 50, 20, 80, 80)



}
}
\seealso{
Other Shape 2D Primitives: \code{\link{arc}},
  \code{\link{ellipse}}, \code{\link{line}},
  \code{\link{point}}, \code{\link{quad}},
  \code{\link{rect}}
}
